Add ngResource.
To replace RestAngular which apparently is no longer maintained.

Also add a patch for ngResource to fix the problem with trailing slashes
being stripped from URLs.
See https://github.com/angular/angular.js/issues/992#issuecomment-33072888 and
https://github.com/appliedsec/djangular.

Note that this problem is fixed in Angular 1.3 with an extra parameter.
The patch checks the Angular version and acts accordingly.